Move settings into the core package
Resolves #291
